Add hyperlink to github-ribbon
Added a onClick handler forkMe() to the github ribbon that opens the
github repository in a new tab/window.

Also re-ordered the github-ribbon image element at the end.
This ensures the img element is rendered on top of other elements on
the page thus avoiding a click "dead-zone" on the github-ribbon.